STDERR output of jobs now make a warning
This is more consistent:
- ERROR if job fails or max_duration exceeded
- WARNING if job prints to stderr
- INFO for job stdout and other lobster outputs

The caveat is that the single lobster log might get unreadable if there are too
many jobs and a lot of output. A header with the job name would help but it's
not simple to implement. And jobs are free to redirect their output anywhere
they want.